move readme fetching to api package
